Refactor BracketedPatternTest to maintain a parallel directory structure
